backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "cat_house_caffe"
version = "0.0.1"
description = "Backend system for ERP, AI assistant, cafe robots, and analytics"
authors = ["Code-Rice <infocoderice@gmail.com>"]
license = "MIT"
readme = "README.md"

[[tool.poetry.packages]]
include = "src"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
django-filter = "^24.1"
django-extensions = "^3.2.3"
django-phonenumber-field = "^7.2.0"
phonenumbers = "^8.13.34"
django-sendgrid = "^0.4.1"
sendgrid = "^6.11.0"
pillow = "^10.3.0"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
openai = "^1.30.1"
langchain = "^0.2.0"
tqdm = "^4.66.4"
requests = "^2.32.2"
httpx = "^0.27.0"
transformers = "^4.41.1"
torch = "^2.3.0"
huggingface-hub = "^0.23.1"
numpy = "^1.26.4"
pandas = "^2.2.2"
scipy = "^1.13.1"
python-dateutil = "^2.9.0.post0"
PyYAML = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
django-debug-toolbar = "^4.3.0"